+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp/src/Expressions/FuncExpression.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System.Collections.Generic;
+using Microsoft.Generator.CSharp.Primitives;
 
 namespace Microsoft.Generator.CSharp.Expressions
 {
