+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp/src/Primitives/CodeFile.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-namespace Microsoft.Generator.CSharp
+namespace Microsoft.Generator.CSharp.Primitives
 {
     public class CodeFile
     {
