+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp/src/Primitives/CodeWriterDeclaration.cs
@@ -3,7 +3,7 @@
 
 using System;
 
-namespace Microsoft.Generator.CSharp
+namespace Microsoft.Generator.CSharp.Primitives
 {
     public sealed class CodeWriterDeclaration
     {
