+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp/src/Primitives/TypeProviderWriter.cs
@@ -2,10 +2,9 @@
 // Licensed under the MIT License.
 
 using System.Linq;
-using Microsoft.Generator.CSharp.Primitives;
 using Microsoft.Generator.CSharp.Providers;
 
-namespace Microsoft.Generator.CSharp
+namespace Microsoft.Generator.CSharp.Primitives
 {
     public class TypeProviderWriter
     {
