+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp/src/Writers/DiagnosticScope.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
+using Microsoft.Generator.CSharp.Primitives;
 
 namespace Microsoft.Generator.CSharp
 {
