+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp/test/Mocks/MockCodeWriterDeclaration.cs
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Microsoft.Generator.CSharp.Primitives;
+
 namespace Microsoft.Generator.CSharp.Tests
 {
     /// <summary>
