+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp/test/Statements/UsingScopeStatementTests.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System.IO;
+using Microsoft.Generator.CSharp.Primitives;
 using Microsoft.Generator.CSharp.Statements;
 using NUnit.Framework;
 using static Microsoft.Generator.CSharp.Snippets.Snippet;
