+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp/test/Writers/CodeScopeTests.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Microsoft.Generator.CSharp.Primitives;
 using NUnit.Framework;
 
 namespace Microsoft.Generator.CSharp.Tests.Writers
